backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
rasterio==1.3.9
opencv-python==4.8.1.78
pillow==10.0.1
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
sentinelhub==3.9.0
scikit-image==0.21.0
matplotlib==3.7.2
